CR comments: Move Terraform specific drift into its own file
